Extract Discord message-turn fakes into tests/support/discord_turn_fakes.py
Move 19 reusable fake classes and builder functions out of the giant
discord_message_turns_support.py spec monolith into a dedicated helper
module, shrinking the support file by ~650 lines.
